Upgrade to upstream v5 major (#1057)
Upgrade to upstream v5 major version. Had to do some interventions here
since upstream changes quite a bit and have completely moved to the PF
for all resources.

- Update the shim to expose the PF TF provider.
- Remove the muxer, since the TF provider is only PF now.
- Remap the IDs for a whole lot of resources which are missing the
field. In Pulumi a string-typed ID field is mandatory, while in PF it
isn't.
- Remove a whole lot of customizations from resources.go for resources
which no longer exist or are now handled by the general ID remapping.
- A Pulumi-side state migration for the `Zone` resource. It had some
breaking changes which would force a replacement.
- Alias all of the renamed resources to their old names.

fixes https://github.com/pulumi/pulumi-cloudflare/issues/1052

---------

+++ sdk/dotnet/Inputs/RulesetRuleActionParametersFromValueGetArgs.cs
@@ -13,19 +13,20 @@ namespace Pulumi.Cloudflare.Inputs
     public sealed class RulesetRuleActionParametersFromValueGetArgs : global::Pulumi.ResourceArgs
     {
         /// <summary>
-        /// Preserve query string for redirect URL.
+        /// Keep the query string of the original request.
         /// </summary>
         [Input("preserveQueryString")]
         public Input<bool>? PreserveQueryString { get; set; }
 
         /// <summary>
-        /// Status code for redirect.
+        /// The status code to be used for the redirect.
+        /// Available values: 301, 302, 303, 307, 308.
         /// </summary>
         [Input("statusCode")]
-        public Input<int>? StatusCode { get; set; }
+        public Input<double>? StatusCode { get; set; }
 
         /// <summary>
-        /// Target URL for redirect.
+        /// The URL to redirect the request to.
         /// </summary>
         [Input("targetUrl")]
         public Input<Inputs.RulesetRuleActionParametersFromValueTargetUrlGetArgs>? TargetUrl { get; set; }
